ChangeLogTag: 2011-10-23    Angelo Corsaro <angelo@icorsaro.net>

+++ src/hpp/dds/core/status/State.hpp
@@ -51,7 +51,12 @@ public:
     }
 
 private:
-    SampleRejectedState(uint32_t s) : MaskType(s) { }
+    // @TODO
+    // -- This Ctor should be fixed as currently there is this 
+    // -- cast only to avoid an error when compiling with the  MS vC++ compiler
+    SampleRejectedState(uint32_t s) 
+        : MaskType((unsigned long long)s) 
+    { } 
 
 };
